+++ project/arduino/smart_waste_disposal/include/level_monitor_task.h
@@ -5,7 +5,7 @@
 
 class LevelMonitorTask: public Task {
     public:
-        void step(long schedPeriod);
+        void step(int schedPeriod);
 };
 
 #endif
