+++ project/arduino/smart_waste_disposal/include/system_tracking_task.h
@@ -5,7 +5,7 @@
 
 class SystemTrackingTask: public Task {
     public:
-        void step(long schedPeriod);
+        void step(int schedPeriod);
 };
 
 #endif
